backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.7.0
pandas>=2.0.0
numpy>=1.26.0
pdfplumber>=0.11.0
openpyxl>=3.1.0
python-multipart>=0.0.9
requests>=2.31.0
pyyaml>=6.0.0
pulp>=2.8.0
